Create upload image component
