+++ frontend/next.config.mjs
@@ -1,6 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-    distDir: "build",
-};
+const nextConfig = {};
 
 export default nextConfig;
