+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "app",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,13 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "bootstrap": "^5.3.3",
-    "next": "14.1.0",
     "react": "^18",
-    "react-dom": "^18"
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.0.1",
-    "postcss": "^8"
+    "react-dom": "^18",
+    "next": "14.1.4"
   }
 }
